update use of plural in base resolver
